requirements.txt
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.6.0
attrs==21.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.6
distro==1.7.0
Django==4.1.6
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0
eyeD3==0.8.10
filelock==3.9.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
grpcio==1.30.2
gunicorn==20.1.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
more-itertools==8.10.0
mpmath==0.0.0
nemo-emblems==5.4.1
netaddr==0.8.0
netifaces==0.11.0
nodejs==0.1.1
numpy==1.21.5
oauthlib==3.2.0
onboard==1.4.1
optional-django==0.1.0
packaging==21.3
PAM==0.4.2
pexpect==4.8.0
Pillow==9.4.0
platformdirs==2.6.2
ply==3.11
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-magic==0.4.24
python-xapp==2.2.2
python-xlib==0.29
pythran==0.10.0
pytz==2022.7
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
scipy==1.8.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
sympy==1.9
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unicodedata2==14.0.0
Unidecode==1.3.3
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.17.1
wadllib==1.3.6
webencodings==0.5.1
whitenoise==6.4.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
